requirements.txt
# Core Selenium stack
selenium==4.18.1
webdriver-manager==4.0.1

# Test runner
pytest==8.1.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==14.0

# Reporting
allure-pytest==2.13.5

# Utilities
python-dotenv==1.0.1
Pillow>=10.3.0
requests==2.31.0
lxml>=4.9.0
